Two more cleanup fixes to the cache.  Don't call destroyDecodedData in the BitmapImage destructor, since         clearing the image source and calling setData on it again causes it to do an extra copy of the encoded         data.  Since we're about to be destroyed this is just wasteful.
        When the cache prunes, don't allow it to destroy the decoded data of an image that is still actively loading,
        since we've established that ImageIO can actually crash if you yank the rug out from under it like that.

        Reviewed by ggaren

        * loader/Cache.cpp:
        (WebCore::Cache::prune):
        * platform/graphics/BitmapImage.cpp:
        (WebCore::BitmapImage::~BitmapImage):



Canonical link: https://commits.webkit.org/16966@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@20169 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/loader/Cache.cpp
@@ -138,7 +138,7 @@ void Cache::prune()
         CachedResource* current = m_lruLists[i].m_tail;
         while (current) {
             CachedResource* prev = current->m_prevInLRUList;
-            if (!current->referenced()) {
+            if (!current->referenced() && current->isLoaded()) {
                 // Go ahead and destroy our decoded data.
                 current->destroyDecodedData();
